+++ packages/vite/src/buildStreamingFeServer.ts
@@ -9,24 +9,21 @@ import { transformWithBabel } from '@redwoodjs/internal/dist/build/babel/api'
 import { buildWeb } from '@redwoodjs/internal/dist/build/web'
 import { findRouteHooksSrc } from '@redwoodjs/internal/dist/files'
 import { getProjectRoutes } from '@redwoodjs/internal/dist/routes'
-import { getAppRouteHook, getConfig, getPaths } from '@redwoodjs/project-config'
+import { getAppRouteHook, getPaths } from '@redwoodjs/project-config'
 
-import { buildRscFeServer } from './buildRscFeServer'
 import { RWRouteManifest } from './types'
 
-export interface BuildOptions {
+interface BuildOptions {
   verbose?: boolean
 }
 
 export const buildFeServer = async ({ verbose }: BuildOptions) => {
   const rwPaths = getPaths()
-  const rwConfig = getConfig()
-  const viteConfigPath = rwPaths.web.viteConfig
+  const viteConfig = rwPaths.web.viteConfig
 
-  if (!viteConfigPath) {
+  if (!viteConfig) {
     throw new Error(
-      'Vite config not found. You need to setup your project with Vite ' +
-        'using `yarn rw setup vite`'
+      'Vite config not found. You need to setup your project with Vite using `yarn rw setup vite`'
     )
   }
 
@@ -38,34 +35,17 @@ export const buildFeServer = async ({ verbose }: BuildOptions) => {
     )
   }
 
-  if (rwConfig.experimental?.rsc?.enabled) {
-    if (!rwPaths.web.entries) {
-      throw new Error('RSC entries file not found')
-    }
-
-    return await buildRscFeServer({
-      viteConfigPath,
-      webSrc: rwPaths.web.src,
-      webHtml: rwPaths.web.html,
-      entries: rwPaths.web.entries,
-      webDist: rwPaths.web.dist,
-      webDistServer: rwPaths.web.distServer,
-      webDistEntries: rwPaths.web.distServerEntries,
-      webRouteManifest: rwPaths.web.routeManifest,
-    })
-  }
-
   // Step 1A: Generate the client bundle
   await buildWeb({ verbose })
 
   // TODO (STREAMING) When Streaming is released Vite will be the only bundler,
   // so we can switch to a regular import
   // @NOTE: Using dynamic import, because vite is still opt-in
-  const { build: viteBuild } = await import('vite')
+  const { build } = await import('vite')
 
   // Step 1B: Generate the server output
-  await viteBuild({
-    configFile: viteConfigPath,
+  await build({
+    configFile: viteConfig,
     build: {
       // Because we configure the root to be web/src, we need to go up one level
       outDir: rwPaths.web.distServer,
